requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.5.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dotenv==0.9.9
ecdsa==0.19.0
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.64.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose[cryptography]
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
